+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.1.2] - 2024-08-04
+### Bugfix/Refactoring
+* Fixed error in Invite Friends method
+* Null checks now done inside VKParams class
 
 ## [2.1.1] - 2024-08-04
 ### Bugfix/Refactoring